#613: tempered with updates becomes the default
