+++ src/lbaf/Execution/lbsCriterionBase.py
@@ -89,11 +89,8 @@ class CriterionBase:
         """Produce the necessary concrete criterion."""
 
         # Load up available criteria
-        # pylint:disable=W0641:possibly-unused-variable,C0415:import-outside-toplevel
         from .lbsTemperedCriterion import TemperedCriterion
-        from .lbsTemperedWithUpdatesCriterion import TemperedWithUpdatesCriterion
         from .lbsStrictLocalizingCriterion import StrictLocalizingCriterion
-        # pylint:enable=W0641:possibly-unused-variable,C0415:import-outside-toplevel
 
         # Ensure that criterion name is valid
         try:
